refactor(ui): rename page-header folder to page
The folder exports the whole page layout family — PageContainer,
PageBody, PageHeader, SecondaryNav — not just the header. Also widens
the page gutters to px-6 and tightens secondary nav spacing.
