+++ web/internal/ui/src/components/page/page-body.tsx
@@ -6,7 +6,7 @@ function PageBody({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       className={cn(
-        "mx-auto w-full max-w-7xl px-4",
+        "mx-auto w-full max-w-7xl px-6",
         "group-data-[width=full]/page:max-w-none group-data-[width=full]/page:px-0",
         className,
       )}
