+++ web/internal/ui/src/components/page/page-header.tsx
@@ -5,7 +5,7 @@ function PageHeader({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       className={cn(
-        "mx-auto flex w-full max-w-7xl items-start justify-between gap-3 px-4 pt-6",
+        "mx-auto flex w-full max-w-7xl items-start justify-between gap-3 px-6 pt-6",
         "group-data-[width=full]/page:max-w-none group-data-[width=full]/page:border-b group-data-[width=full]/page:border-grayA-4 group-data-[width=full]/page:pb-4 group-data-[width=full]/page:pt-4",
         className,
       )}
